Refactor: move auxDump helper to own file
